==> Notebooks/01 - Process Raw Ride-Hailing Data.ipynb <==
import pandas as pd
import numpy as np
import sys 
import datetime
import os
import matplotlib as plt

%matplotlib inline

agg = {'Fare':'mean','Trip Total':'mean','Trip Miles':'sum','Trips Pooled':'sum','PRIVATE_TRIPS':'sum','SHARED_TRIPS':'sum','TRIPS':'sum'}

def clean_float_cols(x):
    x = x.replace(',','')
    x = float(x)
    
    return x

chunk = pd.read_csv('C:/Workspace/TNC-Demand-Model/Inputs/Chicago Ride-Hailing/Transportation_Network_Providers_-_Trips.csv', nrows = 100000, parse_dates = ['Trip Start Timestamp','Trip End Timestamp'], infer_datetime_format = True)

chunk.head()

chunk.columns


#rename columns and drop unnecessary columns
chunk['TRAVEL_TIME_MINUTES'] = chunk['Trip Seconds']/60
chunk['TRIP_LENGTH_MILES'] = chunk['Trip Miles']
chunk['ORIGIN'] = chunk['Pickup Census Tract']
chunk['DESTINATION'] = chunk['Dropoff Census Tract']
chunk['ORIGIN_COM_AREA'] = chunk['Pickup Community Area']
chunk['DEST_COM_AREA'] = chunk['Dropoff Community Area']
chunk['FARE'] = chunk['Fare']
chunk['TIP'] = chunk['Tip']
chunk['ADDITIONAL_CHARGES'] = chunk['Additional Charges']
chunk['TOTAL_COST'] = chunk['Trip Total']
chunk['SHARED_FLAGGER'] = np.where(chunk['Shared Trip Authorized'], 1,0)
chunk['NUM_TRIPS_POOLED'] = chunk['Trips Pooled']
chunk['SPEED'] = chunk['TRIP_LENGTH_MILES']/(chunk['TRAVEL_TIME_MINUTES']/60)
chunk['TRIP_ID'] = chunk['Trip ID']
chunk['TRIP_START_TIME'] = chunk['Trip Start Timestamp']
chunk['TRIP_END_TIME'] = chunk['Trip End Timestamp']



chunk = chunk.drop(['Pickup Census Tract','Dropoff Census Tract','Trip Seconds','Trip Miles','Pickup Community Area','Dropoff Community Area','Fare','Tip','Additional Charges','Trip Total','Shared Trip Authorized','Trips Pooled','Trip ID','Trip Start Timestamp','Trip End Timestamp', 'Pickup Centroid Latitude', 'Pickup Centroid Longitude','Pickup Centroid Location', 'Dropoff Centroid Latitude', 'Dropoff Centroid Longitude', 'Dropoff Centroid Location',],axis = 1)


def assign_flaggers(row):
    if (~np.isnan(row['ORIGIN'])) & (~np.isnan(row['DESTINATION'])) & (~np.isnan(row['ORIGIN_COM_AREA'])) & (~np.isnan(row['DEST_COM_AREA'])):
        row['INTERNAL_UNSUPPRESSED_FLAGGER'] = 1
        row['INTERNAL_SUPPPRESSED_FLAGGER'] = 0
        row['EXTERNAL_FLAGGER'] = 0
        row['INTERNAL_EXTERNAL_FLAGGER'] = 0

    elif (np.isnan(row['ORIGIN'])) & (np.isnan(row['DESTINATION'])) & (~np.isnan(row['ORIGIN_COM_AREA'])) & (~np.isnan(row['DEST_COM_AREA'])):
        row['INTERNAL_UNSUPPRESSED_FLAGGER'] = 0
        row['INTERNAL_SUPPPRESSED_FLAGGER'] = 1
        row['EXTERNAL_FLAGGER'] = 0
        row['INTERNAL_EXTERNAL_FLAGGER'] = 0

    elif (~np.isnan(row['ORIGIN'])) & (np.isnan(row['DESTINATION'])) & (~np.isnan(row['ORIGIN_COM_AREA'])) & (~np.isnan(row['DEST_COM_AREA'])):
        print('Something went wrong!')

    elif (np.isnan(row['ORIGIN'])) & (~np.isnan(row['DESTINATION'])) & (~np.isnan(row['ORIGIN_COM_AREA'])) & (~np.isnan(row['DEST_COM_AREA'])):
        print('Something went wrong!')
        
    elif (np.isnan(row['ORIGIN_COM_AREA'])) & (~np.isnan(row['DEST_COM_AREA'])):
        row['INTERNAL_UNSUPPRESSED_FLAGGER'] = 0
        row['INTERNAL_SUPPPRESSED_FLAGGER'] = 0
        row['EXTERNAL_FLAGGER'] = 0
        row['INTERNAL_EXTERNAL_FLAGGER'] = 1

    elif (~np.isnan(row['ORIGIN_COM_AREA'])) & (np.isnan(row['DEST_COM_AREA'])):
        row['INTERNAL_UNSUPPRESSED_FLAGGER'] = 0
        row['INTERNAL_SUPPPRESSED_FLAGGER'] = 0
        row['EXTERNAL_FLAGGER'] = 0
        row['INTERNAL_EXTERNAL_FLAGGER'] = 1
        
    elif (np.isnan(row['ORIGIN_COM_AREA'])) & (np.isnan(row['DEST_COM_AREA'])):
        row['INTERNAL_UNSUPPRESSED_FLAGGER'] = 0
        row['INTERNAL_SUPPPRESSED_FLAGGER'] = 0
        row['EXTERNAL_FLAGGER'] = 1
        row['INTERNAL_EXTERNAL_FLAGGER'] = 0
        
    else:
        print('Something went wrong!')

    return row

chunk = chunk.apply(lambda row: assign_flaggers(row), axis = 1)

chunk.head()

conflation = pd.read_csv('C:/Workspace/TNC-Demand-Model/Inputs/Chicago Community Areas/Community_Area_to_Census_Tract.csv', index_col =0)
conflation.head()

def process_suppressed(row,conflation):

    if (np.isnan(row['ORIGIN'])) & (np.isnan(row['DESTINATION'])) & (~np.isnan(row['ORIGIN_COM_AREA'])) & (~np.isnan(row['DEST_COM_AREA'])):

        df = pd.DataFrame(data = [row.values] * len(row), columns = row.index)

        df = pd.merge(df[['ORIGIN_COM_AREA','DEST_COM_AREA']],conflation, how = 'left', left_on = 'ORIGIN_COM_AREA', right_on = 'area_num_1',left_index=True, right_index=True)
        df = pd.merge(df,conflation, how = 'left', left_on = 'DEST_COM_AREA', right_on = 'area_num_1', suffixes = ('_ORIGIN','_DEST'),left_index=True, right_index=True)


        df['SCALAR'] = np.random.dirichlet(np.ones(len(df)))
        df = df[df['SCALAR'] == df['SCALAR'].max()]
        
        row['ORIGIN'] = df['GEOID_ORIGIN'].values[0]
        row['DESTINATION'] = df['GEOID_DEST'].values[0]
        
    else:
        row['ORIGIN'] = row['ORIGIN']
        row['DESTINATION'] = row['DESTINATION']
        
    
    return row

#df = row.merge(conflation, how = 'left', left_on = 'ORIGIN_COM_AREA', right_on = ['area_num_1'])
#df = row.merge(df, how = 'left', left_on = 'DEST_COM_AREA', right_on = ['area_num_1'], suffixes = ('_ORIGIN','_DEST'))



df = pd.merge(chunk,conflation, how = 'left', left_on = 'ORIGIN_COM_AREA', right_on = 'area_num_1')

#df[['ORIGIN_COM_AREA','DEST_COM_AREA','GEOID','area_num_1']].head(25)

chunk = chunk.apply(lambda row: process_suppressed(row, conflation), axis = 1)

chunk[chunk['INTERNAL_SUPPPRESSED_FLAGGER'] == 0][['ORIGIN','DESTINATION','INTERNAL_EXTERNAL_FLAGGER','EXTERNAL_FLAGGER']]

chunk.columns

store.close()
