==> Notebooks/03 - Create Empty OD Matrix of Chicago Census Tracts-Copy1.ipynb <==
import pandas as pd
import geopandas as gp

tracts = gp.read_file('C:/Workspace/TNC-Demand-Model/Inputs/Census Shapefiles/Illinois/Chicago Tracts/geo_export_558aad9f-98d8-4dd5-a6b1-c1730155d596.shp')

tods = [1,2,3,4,5]

#df = pd.DataFrame()
#df2 = pd.DataFrame()

#for tract in tracts.geoid10:
   # df['DESTINATION'] = tracts.geoid10
   # df['ORIGIN'] = tract
    #df2 = df2.append(df)

df = pd.DataFrame()
df3 = pd.DataFrame()

years = [2018,2019,2020]

for year in years:
    print('Working on year ' + str(year))

    if year == 2018:
        months = [11,12]
    elif year == 2020:
        months = [1,2]
    else:
        months = [1,2,3,4,5,6,7,8,9,10,11,12]
        
    for month in months:
        print('Working on month ' + str(month))

        for tod in tods:
            df2 = pd.DataFrame()

            print('Working on TOD ' + str(tod))

            for tract in tracts.geoid10:
                df['DESTINATION'] = tracts.geoid10
                df['ORIGIN'] = tract
                df2 = df2.append(df)

            df2['MONTH'] = month
            df2['YEAR'] = year
            df2['TOD'] = tod
            #df2.to_csv('C:/Workspace/TNC-Demand-Model/Outputs/Destination Choice Model Matricies/Empty Matricies/Empty_Chicago_Matrix_' + str(year) + '_' + str(month) + '_' + str(tod) + '.csv')
           
            df3 = df3.append(df2)
df3.to_csv('C:/Workspace/TNC-Demand-Model/Inputs/Chicago Ride-Hailing/Empty_Chicago_Matrix.csv')
print('Complete!')

len(df3)

df3.ORIGIN.value_counts()

#df3.to_csv('C:/Workspace/TNC-Demand-Model/Outputs/Empty_Chicago_Matrix.csv')
