==> Notebooks/04 -Create Census Tract Centroids and Conflation File for Suppressed Trips.ipynb <==
import geopandas as gp
import pandas as pd
import numpy as np

tracts = gp.read_file('C:/Workspace/TNC-Demand-Model/Inputs/Census Shapefiles/Chicago Tracts/geo_export_558aad9f-98d8-4dd5-a6b1-c1730155d596.shp')

origins = pd.DataFrame()
origins['X'] = tracts.centroid.x
origins['Y'] = tracts.centroid.y

origins['GEOID'] = tracts.geoid10

#had to change the centroid of this census tract because the true centroid is located in the middle of a schol and it messes up 
#the transit travel time calculation

origins.loc[origins['GEOID'] == '17031843700', 'X'] = -87.69079
origins.loc[origins['GEOID'] == '17031843700', 'Y'] = 41.94482

#had to change the centroid of this census tract because the true centroid is located in the middle of a waste water treatment 
#plant and it messes up the transit travel time calculation

origins.loc[origins['GEOID'] == '17031540102', 'X'] = -87.60610
origins.loc[origins['GEOID'] == '17031540102', 'Y'] = 41.65970


origins.to_csv('C:/Workspace/TNC-Demand-Model/otp/points.csv')

centroids = gp.GeoDataFrame(tracts.centroid)
centroids['GEOID'] = tracts.geoid10
centroids.columns = ['geometry','GEOID']

com_area = gp.read_file('C:/Workspace/TNC-Demand-Model/Inputs/Chicago Community Areas/geo_export_d8da94d2-4ef1-4ee6-9fff-7bb20d451fe2.shp')

com_area.head()

conflate = gp.sjoin(com_area[['area_num_1','geometry']], centroids[['GEOID','geometry']], how = 'left', op = 'contains' )

conflate.head()

conflate = conflate.drop(columns = ['geometry', 'index_right'])

add_ons = {'area_num_1':[ 53, 75, 43,43,43, 39, 41, 41, 41, 36, 8, 76, 76, 10, 1, 8], 'GEOID':[17031821402, 17031823304, 17031431400, 17031430700, 17031430102, 17031390700, 17031411000, 17031410900, 17031410100, 17031836400,17031081403, 17031770602, 17031770902, 17031810400, 17031010400, 17031081202 ]} 

add_on_df = pd.DataFrame(add_ons)

add_on_df.head()

conflate = conflate.append(add_on_df)

conflate.to_csv('C:/Workspace/TNC-Demand-Model/Inputs/Chicago Community Areas/Community_Area_to_Census_Tract.csv')
