==> Notebooks/05B - Create Monthly 2018-2020 Ride-Hailing Data.ipynb <==
import pandas as pd
import numpy as np
import sys 
import datetime
import os
import matplotlib as plt
import seaborn as sns

%matplotlib inline

# Process the Weekday data

## Averages data for all of 2019

* Average weekday non-pooled trip totals 
* Average weekday pooled trip totals
* Average weekday fares
* Average weekday travel time


tods = [1,2,3,4,5]
years = [2018,2019,2020]

agg = { 'Trip Seconds':'mean',
       'Trip Miles':'mean','Fare':'mean',
       'Tip':'mean', 'Additional Charges':'mean', 'Trip Total':'mean', 'PRIVATE_TRIPS':'sum','SHARED_TRIPS':'sum',
       'Trips Pooled':'sum'}


df_all = pd.DataFrame()
df_raw = pd.DataFrame()

for year in years:
    print('Working on year ' + str(year))
    for tod in tods:
        print('Working on tod ' + str(tod))
        df = pd.read_hdf('C:/Workspace/TNC-Demand-Model/Inputs/Chicago Ride-Hailing/Old Files/Chicago_TNC_Trips_20_Incomplete.H5', where = 'YEAR == ' + str(year), key = 'Weekday_' + str(tod))

        #df = df[df['MONTH'].isin([1,2])]
        #df = df[df['YEAR']==2019]

        print(df['Trip Seconds'].max())
        print(df['Trip Miles'].max())

        df = df.dropna(subset = ['Pickup Census Tract', 'Dropoff Census Tract'])
        df['DAY'] = df['Trip Start Timestamp'].dt.day
        df['PRIVATE_TRIPS'] = np.where(df['Shared Trip Authorized'] == False, 1, 0)
        df['SHARED_TRIPS'] = np.where(df['Shared Trip Authorized'] == True, 1, 0)  

        ## the filters are higher than the longest trip because the trip could be pooled
        # filter out the trips that are longer than 50 miles becuase the longest possible trip is 35 miles going from O'Hare airport to south east corner
        df2 = df[df['Trip Miles'] <= 50]

        # filter out the trips that are longer than 2 hours becuase the longest possible trip is 1 hours going from O'Hare airport to south east corner
        df2 = df2[df2['Trip Seconds'] <= 7200]

        print(df2['Trip Seconds'].max())
        print(df2['Trip Miles'].max())

        #census tract XXX replaces the trips assigned to census tract XXX because they are likely misasigned. Census tract XXX contains
        #contains O'Hare airport and census tract is adjacent. The trips assigned to census tract XXX are likely trips from the airport.
        df2.loc[df2['Pickup Census Tract'] == 17031770700, 'Pickup Census Tract'] = 17031980000
        df2.loc[df2['Dropoff Census Tract'] == 17031770700, 'Dropoff Census Tract'] = 17031980000


        #trips to census tract XXX are likely misasigned but it is not clear which census tract they should be assigned to.
        #there is only one trip, so it is droped
        df2 = df2[df2['Pickup Census Tract'] != 17031810502]
        df2 = df2[df2['Dropoff Census Tract'] != 17031810502]

       # if df2['MONTH'] == [1,2,3]:
        #    df2['QUARTER'] = 1

        #elif df2['MONTH'] == [4,5,6]:
         #   df2['QUARTER'] = 2

        #elif df2['MONTH'] == [7,8,9]:
         #   df2['QUARTER'] = 3

        #elif df2['MONTTH'] == [10,11,12]:
         #   df2['QUARTER'] = 4

        df2 = df2.groupby(by= ['Pickup Census Tract','Dropoff Census Tract','YEAR', 'MONTH','DAY'], as_index =False).agg(agg)
        df2 = df2.groupby(by= ['Pickup Census Tract','Dropoff Census Tract','YEAR','MONTH'], as_index =False).mean()

        df['TOD'] = tod
        df2['TOD'] = tod

        df_raw = df_raw.append(df)
        df_all = df_all.append(df2)

## Add in the Suppressed Trips

sup_trips = pd.read_csv('C:/Workspace/TNC-Demand-Model/Inputs/Chicago Ride-Hailing/2019 Suppressed Trips.csv')

df_all.head()

sup_trips.head()

df_final = df_all.merge(sup_trips[['GEOID_PICKUP', 'GEOID_DROPOFF', 'SCALED_SUP_PRIVATE_TRIPS', 'SCALED_SUP_SHARED_TRIPS', 'TOD']], how = 'left', left_on = ['Pickup Census Tract','Dropoff Census Tract', 'TOD'],right_on = ['GEOID_PICKUP', 'GEOID_DROPOFF', 'TOD'])

df_final.head()

df_final['SHARED_TRIPS'] = df_final['SHARED_TRIPS'] + df_final['SCALED_SUP_SHARED_TRIPS']
df_final['PRIVATE_TRIPS'] = df_final['PRIVATE_TRIPS'] + df_final['SCALED_SUP_PRIVATE_TRIPS']
df_final['ALL_TRIPS'] = df_final['PRIVATE_TRIPS'] + df_final['SHARED_TRIPS']

# Merge RH Data to Empty Chicago OD Matrix

empty = pd.read_csv('Inputs/Chicago Ride-Hailing/Empty_Chicago_Matrix.csv')

df_final.head()

# this is a way to convert the chicago data file tract column from a float to a int.. save for later

#test = df_final['Pickup Census Tract'].astype(str)

#int(test[0][:-2])

empty.DESTINATION = empty.DESTINATION.astype(float)
empty.ORIGIN = empty.ORIGIN.astype(float)

rh_final = empty.merge(df_final, how = 'left', left_on = ['ORIGIN', 'DESTINATION','TOD'], right_on = ['Pickup Census Tract', 'Dropoff Census Tract', 'TOD'])

rh_final.head()

len(rh_final)

len(df_final)

rh_final = rh_final.fillna(0)

rh_final.ALL_TRIPS.value_counts()

rh_final.to_csv('Outputs/2019_Weekday_Yearly_Ridehail_TOD.csv')

rh_final_day = rh_final[['ORIGIN','DESTINATION','ALL_TRIPS']].groupby(by = ['ORIGIN','DESTINATION'], as_index = False).sum()

df_raw.to_csv('Inputs/Chicago Ride-Hailing/Raw_Trip_Records_No_Suppressed.csv')

df_all.to_csv('Inputs/Chicago Ride-Hailing/Monthly_Trip_Records_No_Suppressed.csv')

rh_final.TOD.value_counts()

# Data Accuracy Check

df = pd.read_hdf('C:/Workspace/TNC-Demand-Model/Inputs/Chicago Ride-Hailing/Chicago_TNC_Trips_20.H5', key = 'Weekday_' + str(3))

jan_feb = df[df['MONTH'].isin([1,2])]

miles_over_50 = jan_feb[jan_feb['Trip Miles'] > 50]

len(miles_over_50)

miles_over_50.to_csv('Trips_Over_50_Miles.csv')

fare_0 = jan_feb[jan_feb['Fare'] == 0]

fare_0.to_csv('Fare_0_trips.csv')

jan_feb[(jan_feb['Trip Start Timestamp'] == '2019-01-15 08:30:00')&(jan_feb['Pickup Census Tract'] == 17031839700)&(jan_feb['Dropoff Census Tract'] == 17031320100)]

miles_over_50.head()

jan_feb[jan_feb['Trip Total'] == 0]

len(jan_feb[jan_feb['Fare'] == 0])

trips_5hrs_moore = jan_feb[jan_feb['Trip Seconds'] >= 18000]

trips_5hrs_moore.to_csv('trips_over_5hrs.csv')
