==> Notebooks/07 - Process ACS Data.ipynb <==
import pandas as pd 
import numpy as np

year = 2018

acs_dp05 = pd.read_csv('D:/TNC-Demand-Model/Inputs/ACS/' + str(year) + '/DP05/DP05.csv')

acs_dp05.head()

acs_dp05[['TRACT', 'COUNTY', 'STATE']] = acs_dp05.NAME.str.split(',', expand=True)

acs_dp05 = acs_dp05[acs_dp05['COUNTY'] == ' Cook County']

acs_dp05.head()

acs_dp05['CENSUS_TRACT'] = acs_dp05.apply(lambda row: row['GEO_ID'][9:], axis = 1)

acs_dp05.CENSUS_TRACT.head()

rename = ['TOTAL_POP', 'TOTAL_MALE', 'TOTAL_FEMALE', 'AGE_5U', 'AGE_5_9', 'AGE+10_14', 'AGE_15_19', 'AGE_20_24', 'AGE_25_34', 'AGE_35_44', 'AGE_45_54', 'AGE_55_59', 'AGE_60_64', 'AGE_65_74', 'AGE_75_84', 'AGE_85P', 'MEDIAN_AGE', 'CENSUS_TRACT']

keep = ['DP05_0001E', 'DP05_0002E', 'DP05_0003E', 'DP05_0005E', 'DP05_0006E', 'DP05_0007E','DP05_0008E', 'DP05_0009E', 'DP05_0010E', 'DP05_0011E', 'DP05_0012E', 'DP05_0013E', 'DP05_0014E', 'DP05_0015E', 'DP05_0016E', 'DP05_0017E', 'DP05_0018E', 'CENSUS_TRACT']

acs_dp05 = acs_dp05[keep]

acs_dp05.columns = rename

acs_dp05.head()

acs_s1101 = pd.read_csv('C:/Workspace/TNC-Demand-Model/Inputs/ACS/' + str(year) + '/S1101/S1101.csv')

acs_s1101.head()

acs_s1101[['TRACT', 'COUNTY', 'STATE']] = acs_s1101.NAME.str.split(',', expand=True)

acs_s1101 = acs_s1101[acs_s1101['COUNTY'] == ' Cook County']

acs_s1101['CENSUS_TRACT'] = acs_s1101.apply(lambda row: row['GEO_ID'][9:], axis = 1)

keep = ['S1101_C01_001E', 'S1101_C01_002E', 'S1101_C01_004E', 'CENSUS_TRACT']

rename = ['TOTAL_HHLDS', 'AVG_HHLD_SIZE', 'AVG_FAMILY_SIZE', 'CENSUS_TRACT']

acs_s1101 = acs_s1101[keep]
acs_s1101.columns = rename

acs_s1501 = pd.read_csv('C:/Workspace/TNC-Demand-Model/Inputs/ACS/' + str(year) + '/S1501/S1501.csv')

acs_s1501.head()

acs_s1501[['TRACT', 'COUNTY', 'STATE']] = acs_s1501.NAME.str.split(',', expand=True)

acs_s1501 = acs_s1501[acs_s1501['COUNTY'] == ' Cook County']

acs_s1501['CENSUS_TRACT'] = acs_s1501.apply(lambda row: row['GEO_ID'][9:], axis = 1)

keep = ['S1501_C01_001E', 'S1501_C01_002E', 'S1501_C01_003E', 'S1501_C01_004E', 'S1501_C01_005E', 'S1501_C01_006E', 'S1501_C01_007E', 'S1501_C01_008E', 'S1501_C01_009E', 'S1501_C01_010E', 'S1501_C01_011E', 'S1501_C01_012E', 'S1501_C01_013E', 'S1501_C01_014E', 'S1501_C01_015E','CENSUS_TRACT']

rename = ['TOTAL_POP_18_24', 'TOTAL_POP_18_24_NO_HIGH_SCHOOL', 'TOTAL_POP_18_24_HIGH_SCHOOL', 'TOTAL_POP_18_24_SOME_COLLEGE', 'TOTAL_POP_18_24_SOME_BACHELORS_HIGHER', 'TOTAL_POP_25P', 'TOTAL_POP_25P_LESS_THAN_9TH', 'TOTAL_POP_25P_9TH_TO_12TH', 'TOTAL_POP_25P_HIGH_SCHOOL', 'TOTAL_POP_25P_SOME_COLLEGE', 'TOTAL_POP_25P_ASSOCIATES', 'TOTAL_POP_25P_BACHELORS', 'TOTAL_POP_25P_GRADUATE', 'TOTAL_POP_25P_HIGH_SCHOOL_HIGHER', 'TOTAL_POP_25P_BACHELORS_HIGHER','CENSUS_TRACT']

acs_s1501 = acs_s1501[keep]
acs_s1501.columns = rename

acs_s1901 = pd.read_csv('C:/Workspace/TNC-Demand-Model/Inputs/ACS/' + str(year) + '/S1901/S1901.csv')

acs_s1901.head()

acs_s1901[['TRACT', 'COUNTY', 'STATE']] = acs_s1901.NAME.str.split(',', expand=True)

acs_s1901 = acs_s1901[acs_s1901['COUNTY'] == ' Cook County']

acs_s1901['CENSUS_TRACT'] = acs_s1901.apply(lambda row: row['GEO_ID'][9:], axis = 1)

keep = ['S1901_C01_002E', 'S1901_C01_003E', 'S1901_C01_004E', 'S1901_C01_005E', 'S1901_C01_006E', 'S1901_C01_007E', 'S1901_C01_008E', 'S1901_C01_009E', 'S1901_C01_010E', 'S1901_C01_011E', 'S1901_C01_012E', 'S1901_C01_012E','CENSUS_TRACT']

rename = ['TOTAL_HHLDS_LESS_10K', 'TOTAL_HHLDS_10K_15K', 'TOTAL_HHLDS_15K_25K', 'TOTAL_HHLDS_25K_35K', 'TOTAL_HHLDS_35K_50K', 'TOTAL_HHLDS_50K_75K', 'TOTAL_HHLDS_75K_100K', 'TOTAL_HHLDS_100K_150K', 'TOTAL_HHLDS_150K_200K', 'TOTAL_HHLDS_200KP', 'HHLDS_MEDIAN_INCOME', 'HHLDS_MEAN_INCOME','CENSUS_TRACT']

acs_s1901 = acs_s1901[keep]
acs_s1901.columns = rename

acs_dp04 = pd.read_csv('C:/Workspace/TNC-Demand-Model/Inputs/ACS/' + str(year) + '/DP04/DP04.csv')

acs_dp04[['TRACT', 'COUNTY', 'STATE']] = acs_dp04.NAME.str.split(',', expand=True)

acs_dp04 = acs_dp04[acs_dp04['COUNTY'] == ' Cook County']

acs_dp04['CENSUS_TRACT'] = acs_dp04.apply(lambda row: row['GEO_ID'][9:], axis = 1)

acs_dp04.CENSUS_TRACT.head()

rename = ['TOTAL_HHLDS', 'HHLDS_0_VEH', 'HHLDS_1_VEH', 'HHLDS_2_VEH', 'HHLDS_3P_VEH', 'CENSUS_TRACT']

keep = ['DP04_0057E', 'DP04_0058E', 'DP04_0059E', 'DP04_0060E', 'DP04_0061E', 'CENSUS_TRACT']

acs_dp04 = acs_dp04[keep]

acs_dp04.columns = rename

acs_dp04.head()

#THIS ALL NEEDS TO GO IN THE MAIN SCRIPT

acs_dp04['HHLDS_0_VEH'] = acs_dp04.HHLDS_0_VEH.astype(float)
acs_dp04['HHLDS_1_VEH'] = acs_dp04.HHLDS_1_VEH.astype(float)
acs_dp04['HHLDS_2_VEH'] = acs_dp04.HHLDS_2_VEH.astype(float)
acs_dp04['HHLDS_3P_VEH'] = acs_dp04.HHLDS_3P_VEH.astype(float)
acs_dp04['TOTAL_HHLDS'] = acs_dp04.TOTAL_HHLDS.astype(float)

#THIS ALL NEEDS TO GO IN THE MAIN SCRIPT

def set_avg_veh_per_hhlds(row):
    if row['TOTAL_HHLDS']  == 0:
        veh = 0
        
    else:
        veh = (row['HHLDS_0_VEH']*0 + row['HHLDS_1_VEH']*1 + row['HHLDS_2_VEH']*2 + row['HHLDS_3P_VEH']*3)/row['TOTAL_HHLDS']
        
    return veh

#THIS ALL NEEDS TO GO IN THE MAIN SCRIPT

acs_dp04['AVG_VEH_PER_HHLD'] = acs_dp04.apply(lambda row: set_avg_veh_per_hhlds(row), axis = 1)

acs_dp04.head()

#THIS ALL NEEDS TO GO IN THE MAIN SCRIPT

acs = acs_dp05.merge(acs_s1101, on = 'CENSUS_TRACT', suffixes = ('_DP05', '_S1101'))
acs = acs.merge(acs_s1501, on = 'CENSUS_TRACT', suffixes = ('', '_S1501'))
acs = acs.merge(acs_s1901, on = 'CENSUS_TRACT', suffixes = ('', '_S1901'))
acs = acs.merge(acs_dp04, on = 'CENSUS_TRACT', suffixes = ('', '_DP04'))

acs.head()

#THIS ALL NEEDS TO GO IN THE MAIN SCRIPT

variables = acs.columns.drop('CENSUS_TRACT')

#THIS ALL NEEDS TO GO IN THE MAIN SCRIPT


for var in variables:    
    acs[var] = acs[var].replace('-',np.nan)
    acs[var] = acs[var].replace('250,000+','250000')
    acs[var] = acs[var].astype(float)

#THIS ALL NEEDS TO GO IN THE MAIN SCRIPT


acs.to_csv('C:/Workspace/TNC-Demand-Model/Outputs/ACS_' + str(year) + '.csv')

len(acs)

len(acs_dp05)

len(acs_s1501)

len(acs_s1901)

len(acs_s1101)

len(acs_dp04)
