==> Notebooks/10 - Link Data Into Estimation File.ipynb <==
import pandas as pd
import geopandas as gp
import numpy as np

ridehail = pd.read_csv('D:/TNC-Demand-Model/Outputs/2019_Weekday_Monthly_Ridehail_TOD.csv')

len(ridehail)

agg = {'Trip Seconds':'mean', 'Trip Miles':'mean', 'Fare':'mean', 'Tip':'mean',
       'Additional Charges':'mean', 'Trip Total':'mean', 'PRIVATE_TRIPS':'sum', 'SHARED_TRIPS':'sum',
       'SCALED_SUP_PRIVATE_TRIPS':'sum', 'SCALED_SUP_SHARED_TRIPS':'sum', 'ALL_TRIPS':'sum'}

ridehail_daily = ridehail.groupby(by = ['YEAR','MONTH','ORIGIN','DESTINATION'], as_index = False).agg(agg)

ridehail = pd.DataFrame()

acs = pd.read_csv('D:/TNC-Demand-Model/Outputs/Chicago_ACS_Data.csv')


acs_daily = acs.groupby(by = ['YEAR','MONTH','CENSUS_TRACT'], as_index = False).mean()

acs_daily = acs_daily.drop(['Unnamed: 0','TOD'], axis = 1)

acs = pd.DataFrame()

lehd_od = pd.read_csv('D:/TNC-Demand-Model/Outputs/CHI_LEHD_OD.csv')
lehd_rac = pd.read_csv('D:/TNC-Demand-Model/Outputs/CHI_RAC.csv')
lehd_wac = pd.read_csv('D:/TNC-Demand-Model/Outputs/CHI_WAC.csv')

lehd_od = lehd_od.drop(['Unnamed: 0'], axis = 1)
lehd_rac = lehd_rac.drop(['Unnamed: 0'], axis = 1)
lehd_wac = lehd_wac.drop(['Unnamed: 0'], axis = 1)

otp = pd.read_csv('D:/TNC-Demand-Model/Outputs/OTP_Transit_Travel_Times.csv')

otp_daily = otp.groupby(by = ['YEAR', 'MONTH','origin','destination'],as_index = False).mean()

otp_daily = otp_daily.drop(['Unnamed: 0','year','TOD'], axis = 1)

otp = pd.DataFrame()

len(acs_daily)

len(lehd_od.TRACTCE10_PICKUP.unique())

est = ridehail_daily.merge(acs_daily, how = 'left', left_on = ['YEAR', 'MONTH', 'ORIGIN'], right_on = ['YEAR', 'MONTH', 'CENSUS_TRACT'], suffixes = ('','_ORIGIN'))
est = est.merge(acs_daily, how = 'left', left_on = ['YEAR', 'MONTH', 'DESTINATION'], right_on = ['YEAR', 'MONTH', 'CENSUS_TRACT'], suffixes = ('','_DESTINATION'))

est['TRACTCE10_PICKUP'] = est.ORIGIN.astype(str)

len(est.TRACTCE10_PICKUP.unique())

est['TRACTCE10_PICKUP'] = est['TRACTCE10_PICKUP'].apply(lambda x : x[5:11])
est['TRACTCE10_PICKUP'] = est['TRACTCE10_PICKUP'].astype(int)

est['TRACTCE10_DEST'] = est.DESTINATION.astype(str)

est['TRACTCE10_DEST'] = est['TRACTCE10_DEST'].apply(lambda x : x[5:11])
est['TRACTCE10_DEST'] = est['TRACTCE10_DEST'].astype(int)

est = est.merge(lehd_od, how = 'left', on = ['YEAR', 'MONTH','TRACTCE10_PICKUP', 'TRACTCE10_DEST'], validate = 'one_to_one', suffixes = ('', '_LEHD_OD'))

est[np.isnan(est['SI01'])]

lehd_od[(lehd_od['TRACTCE10_PICKUP'] == 10100)&(lehd_od['TRACTCE10_DEST'] == 10400)]

est.SI01.value_counts(dropna = False)

lehd_wac.columns

est['SI01'] = est['SI01'].fillna(0)
est['SI02'] = est['SI02'].fillna(0)
est['SI03'] = est['SI03'].fillna(0)

est = est.merge(lehd_rac, how = 'left', right_on = ['YEAR', 'MONTH','TRACTCE10'], left_on = ['YEAR','MONTH','TRACTCE10_PICKUP'], suffixes = ('', '_LEHD_RAC'))
est = est.merge(lehd_rac, how = 'left', right_on = ['YEAR', 'MONTH','TRACTCE10'], left_on = ['YEAR','MONTH', 'TRACTCE10_DEST'], suffixes = ('_RAC_ORIGIN', '_RAC_DESTINATION'))

est = est.merge(lehd_wac, how = 'left', right_on = ['YEAR', 'MONTH', 'TRACTCE10'], left_on = ['YEAR','MONTH', 'TRACTCE10_PICKUP'], suffixes = ('', '_LEHD_WAC'))
est = est.merge(lehd_wac, how = 'left', right_on = ['YEAR', 'MONTH', 'TRACTCE10'], left_on = ['YEAR','MONTH','TRACTCE10_DEST'], suffixes = ('_WAC_ORIGIN', '_WAC_DESTINATION'))


fill_cols = lehd_wac.columns.drop(['YEAR','MONTH','TRACTCE10'])

for col in fill_cols:
    col = col + '_WAC_ORIGIN'
    est[col] = est[col].fillna(0)

for col in fill_cols:
    col = col + '_WAC_DESTINATION'
    est[col] = est[col].fillna(0)

ridehail_daily = pd.DataFrame()
acs_daily = pd.DataFrame()
acs_daily = pd.DataFrame()
lehd_od = pd.DataFrame()
lehd_rac = pd.DataFrame()
lehd_wac = pd.DataFrame()

otp_daily.info()

est[np.isnan(est['CNS01_WAC_ORIGIN'])]['ORIGIN']

est.info()

otp_daily.info()

otp_daily.head()

otp_daily[(otp_daily['origin'] == 17031010100)&(otp_daily['destination'] == 17031010100)&(otp_daily['MONTH'] == 11)]

est = est.merge(otp_daily, how = 'left', right_on = ['YEAR', 'MONTH', 'origin','destination'], left_on = ['YEAR','MONTH','ORIGIN','DESTINATION'], validate = 'one_to_one', suffixes = ('', '_OTP'))

est.head()

ridehail_daily = pd.DataFrame()
acs = pd.DataFrame()

YEAR
MONTH
ORIGIN
DESTINATION
Trip Seconds
Trip Miles
Fare
Tip
Additional Charges
Trip Total
PRIVATE_TRIPS
SHARED_TRIPS
SCALED_SUP_PRIVATE_TRIPS
SCALED_SUP_SHARED_TRIPS
ALL_TRIPS
CENSUS_TRACT
TOTAL_POP
TOTAL_MALE
TOTAL_FEMALE
AGE_5U
AGE_5_9
AGE+10_14
AGE_15_19
AGE_20_24
AGE_25_34
AGE_35_44
AGE_45_54
AGE_55_59
AGE_60_64
AGE_65_74
AGE_75_84
AGE_85P
MEDIAN_AGE
TOTAL_HHLDS
AVG_HHLD_SIZE
AVG_FAMILY_SIZE
TOTAL_POP_18_24
TOTAL_POP_18_24_NO_HIGH_SCHOOL
TOTAL_POP_18_24_HIGH_SCHOOL
TOTAL_POP_18_24_SOME_COLLEGE
TOTAL_POP_18_24_SOME_BACHELORS_HIGHER
TOTAL_POP_25P
TOTAL_POP_25P_LESS_THAN_9TH
TOTAL_POP_25P_9TH_TO_12TH
TOTAL_POP_25P_HIGH_SCHOOL
TOTAL_POP_25P_SOME_COLLEGE
TOTAL_POP_25P_ASSOCIATES
TOTAL_POP_25P_BACHELORS
TOTAL_POP_25P_GRADUATE
TOTAL_POP_25P_HIGH_SCHOOL_HIGHER
TOTAL_POP_25P_BACHELORS_HIGHER
TOTAL_HHLDS_LESS_10K
TOTAL_HHLDS_10K_15K
TOTAL_HHLDS_15K_25K
TOTAL_HHLDS_25K_35K
TOTAL_HHLDS_35K_50K
TOTAL_HHLDS_50K_75K
TOTAL_HHLDS_75K_100K
TOTAL_HHLDS_100K_150K
TOTAL_HHLDS_150K_200K
TOTAL_HHLDS_200KP
HHLDS_MEDIAN_INCOME
HHLDS_MEAN_INCOME
HHLDS_0_VEH
HHLDS_1_VEH
HHLDS_2_VEH
HHLDS_3P_VEH
AVG_VEH_PER_HHLD
CENSUS_TRACT_DESTINATION
TOTAL_POP_DESTINATION
TOTAL_MALE_DESTINATION
TOTAL_FEMALE_DESTINATION
AGE_5U_DESTINATION
AGE_5_9_DESTINATION
AGE+10_14_DESTINATION
AGE_15_19_DESTINATION
AGE_20_24_DESTINATION
AGE_25_34_DESTINATION
AGE_35_44_DESTINATION
AGE_45_54_DESTINATION
AGE_55_59_DESTINATION
AGE_60_64_DESTINATION
AGE_65_74_DESTINATION
AGE_75_84_DESTINATION
AGE_85P_DESTINATION
MEDIAN_AGE_DESTINATION
TOTAL_HHLDS_DESTINATION
AVG_HHLD_SIZE_DESTINATION
AVG_FAMILY_SIZE_DESTINATION
TOTAL_POP_18_24_DESTINATION
TOTAL_POP_18_24_NO_HIGH_SCHOOL_DESTINATION
TOTAL_POP_18_24_HIGH_SCHOOL_DESTINATION
TOTAL_POP_18_24_SOME_COLLEGE_DESTINATION
TOTAL_POP_18_24_SOME_BACHELORS_HIGHER_DESTINATION
TOTAL_POP_25P_DESTINATION
TOTAL_POP_25P_LESS_THAN_9TH_DESTINATION
TOTAL_POP_25P_9TH_TO_12TH_DESTINATION
TOTAL_POP_25P_HIGH_SCHOOL_DESTINATION
TOTAL_POP_25P_SOME_COLLEGE_DESTINATION
TOTAL_POP_25P_ASSOCIATES_DESTINATION
TOTAL_POP_25P_BACHELORS_DESTINATION
TOTAL_POP_25P_GRADUATE_DESTINATION
TOTAL_POP_25P_HIGH_SCHOOL_HIGHER_DESTINATION
TOTAL_POP_25P_BACHELORS_HIGHER_DESTINATION
TOTAL_HHLDS_LESS_10K_DESTINATION
TOTAL_HHLDS_10K_15K_DESTINATION
TOTAL_HHLDS_15K_25K_DESTINATION
TOTAL_HHLDS_25K_35K_DESTINATION
TOTAL_HHLDS_35K_50K_DESTINATION
TOTAL_HHLDS_50K_75K_DESTINATION
TOTAL_HHLDS_75K_100K_DESTINATION
TOTAL_HHLDS_100K_150K_DESTINATION
TOTAL_HHLDS_150K_200K_DESTINATION
TOTAL_HHLDS_200KP_DESTINATION
HHLDS_MEDIAN_INCOME_DESTINATION
HHLDS_MEAN_INCOME_DESTINATION
HHLDS_0_VEH_DESTINATION
HHLDS_1_VEH_DESTINATION
HHLDS_2_VEH_DESTINATION
HHLDS_3P_VEH_DESTINATION
AVG_VEH_PER_HHLD_DESTINATION
TRACTCE10_PICKUP
TRACTCE10_DEST
SI01
SI02
SI03
TRACTCE10_RAC_ORIGIN
CNS01_RAC_ORIGIN
CNS02_RAC_ORIGIN
CNS03_RAC_ORIGIN
CNS04_RAC_ORIGIN
CNS05_RAC_ORIGIN
CNS06_RAC_ORIGIN
CNS07_RAC_ORIGIN
CNS08_RAC_ORIGIN
CNS09_RAC_ORIGIN
CNS10_RAC_ORIGIN
CNS11_RAC_ORIGIN
CNS12_RAC_ORIGIN
CNS13_RAC_ORIGIN
CNS14_RAC_ORIGIN
CNS15_RAC_ORIGIN
CNS16_RAC_ORIGIN
CNS17_RAC_ORIGIN
CNS18_RAC_ORIGIN
CNS19_RAC_ORIGIN
CNS20_RAC_ORIGIN
TRACTCE10_RAC_DESTINATION
CNS01_RAC_DESTINATION
CNS02_RAC_DESTINATION
CNS03_RAC_DESTINATION
CNS04_RAC_DESTINATION
CNS05_RAC_DESTINATION
CNS06_RAC_DESTINATION
CNS07_RAC_DESTINATION
CNS08_RAC_DESTINATION
CNS09_RAC_DESTINATION
CNS10_RAC_DESTINATION
CNS11_RAC_DESTINATION
CNS12_RAC_DESTINATION
CNS13_RAC_DESTINATION
CNS14_RAC_DESTINATION
CNS15_RAC_DESTINATION
CNS16_RAC_DESTINATION
CNS17_RAC_DESTINATION
CNS18_RAC_DESTINATION
CNS19_RAC_DESTINATION
CNS20_RAC_DESTINATION
TRACTCE10_WAC_ORIGIN
CNS01_WAC_ORIGIN
CNS02_WAC_ORIGIN
CNS03_WAC_ORIGIN
CNS04_WAC_ORIGIN
CNS05_WAC_ORIGIN
CNS06_WAC_ORIGIN
CNS07_WAC_ORIGIN
CNS08_WAC_ORIGIN
CNS09_WAC_ORIGIN
CNS10_WAC_ORIGIN
CNS11_WAC_ORIGIN
CNS12_WAC_ORIGIN
CNS13_WAC_ORIGIN
CNS14_WAC_ORIGIN
CNS15_WAC_ORIGIN
CNS16_WAC_ORIGIN
CNS17_WAC_ORIGIN
CNS18_WAC_ORIGIN
CNS19_WAC_ORIGIN
CNS20_WAC_ORIGIN
TRACTCE10_WAC_DESTINATION
CNS01_WAC_DESTINATION
CNS02_WAC_DESTINATION
CNS03_WAC_DESTINATION
CNS04_WAC_DESTINATION
CNS05_WAC_DESTINATION
CNS06_WAC_DESTINATION
CNS07_WAC_DESTINATION
CNS08_WAC_DESTINATION
CNS09_WAC_DESTINATION
CNS10_WAC_DESTINATION
CNS11_WAC_DESTINATION
CNS12_WAC_DESTINATION
CNS13_WAC_DESTINATION
CNS14_WAC_DESTINATION
CNS15_WAC_DESTINATION
CNS16_WAC_DESTINATION
CNS17_WAC_DESTINATION
CNS18_WAC_DESTINATION
CNS19_WAC_DESTINATION
CNS20_WAC_DESTINATION
origin
destination
walk_distance
travel_time

est['TOTAL_JOBS_ORIGIN'] = est['CNS01_WAC_ORIGIN'] +est['CNS02_WAC_ORIGIN']+est['CNS03_WAC_ORIGIN']+est['CNS04_WAC_ORIGIN']+est['CNS05_WAC_ORIGIN']+est['CNS06_WAC_ORIGIN']+est['CNS07_WAC_ORIGIN']+est['CNS08_WAC_ORIGIN']+est['CNS09_WAC_ORIGIN']+est['CNS10_WAC_ORIGIN']+est['CNS11_WAC_ORIGIN']+est['CNS12_WAC_ORIGIN']+est['CNS13_WAC_ORIGIN']+est['CNS14_WAC_ORIGIN']+est['CNS15_WAC_ORIGIN']+est['CNS16_WAC_ORIGIN']+est['CNS17_WAC_ORIGIN']+est['CNS18_WAC_ORIGIN']+est['CNS19_WAC_ORIGIN']+est['CNS20_WAC_ORIGIN']

est['TOTAL_JOBS_DESTINATION'] = est['CNS01_WAC_DESTINATION'] +est['CNS02_WAC_DESTINATION']+est['CNS03_WAC_DESTINATION']+est['CNS04_WAC_DESTINATION']+est['CNS05_WAC_DESTINATION']+est['CNS06_WAC_DESTINATION']+est['CNS07_WAC_DESTINATION']+est['CNS08_WAC_DESTINATION']+est['CNS09_WAC_DESTINATION']+est['CNS10_WAC_DESTINATION']+est['CNS11_WAC_DESTINATION']+est['CNS12_WAC_DESTINATION']+est['CNS13_WAC_DESTINATION']+est['CNS14_WAC_DESTINATION']+est['CNS15_WAC_DESTINATION']+est['CNS16_WAC_DESTINATION']+est['CNS17_WAC_DESTINATION']+est['CNS18_WAC_DESTINATION']+est['CNS19_WAC_DESTINATION']+est['CNS20_WAC_DESTINATION']

est['TOTAL_WORKERS_ORIGIN'] = est['CNS01_RAC_ORIGIN']+est['CNS02_RAC_ORIGIN']+est['CNS03_RAC_ORIGIN']+est['CNS04_RAC_ORIGIN']+est['CNS05_RAC_ORIGIN']+est['CNS06_RAC_ORIGIN']+est['CNS07_RAC_ORIGIN']+est['CNS08_RAC_ORIGIN']+est['CNS09_RAC_ORIGIN']+est['CNS10_RAC_ORIGIN']+est['CNS11_RAC_ORIGIN']+est['CNS12_RAC_ORIGIN']+est['CNS13_RAC_ORIGIN']+est['CNS14_RAC_ORIGIN']+est['CNS15_RAC_ORIGIN']+est['CNS16_RAC_ORIGIN']+est['CNS17_RAC_ORIGIN']+est['CNS18_RAC_ORIGIN']+est['CNS19_RAC_ORIGIN']+est['CNS20_RAC_ORIGIN']

est['TOTAL_WORKERS_DESTINATION'] = est['CNS01_RAC_DESTINATION']+est['CNS02_RAC_DESTINATION']+est['CNS03_RAC_DESTINATION']+est['CNS04_RAC_DESTINATION']+est['CNS05_RAC_DESTINATION']+est['CNS06_RAC_DESTINATION']+est['CNS07_RAC_DESTINATION']+est['CNS08_RAC_DESTINATION']+est['CNS09_RAC_DESTINATION']+est['CNS10_RAC_DESTINATION']+est['CNS11_RAC_DESTINATION']+est['CNS12_RAC_DESTINATION']+est['CNS13_RAC_DESTINATION']+est['CNS14_RAC_DESTINATION']+est['CNS15_RAC_DESTINATION']+est['CNS16_RAC_DESTINATION']+est['CNS17_RAC_DESTINATION']+est['CNS18_RAC_DESTINATION']+est['CNS19_RAC_DESTINATION']+est['CNS20_RAC_DESTINATION']

for col in est.columns:
    print(col)

est.to_csv('D:/TNC-Demand-Model/Outputs/Chi_Estimation_File.csv')

#rename = ['Pickup Census Tract',
'Dropoff Census Tract',
'MONTH',
'DAY',
'Trip Seconds',
'Trip Miles',
'Fare',
'Tip',
'Additional Charges',
'Trip Total',
'PRIVATE_TRIPS',
'SHARED_TRIPS',
'Trips Pooled',
'TOD',
'TOTAL_POP_ORIGIN',
'TOTAL_MALE_ORIGIN',
'TOTAL_FEMALE_ORIGIN',
'AGE_5U_ORIGIN',
'AGE_5_9_ORIGIN',
'AGE_10_14_ORIGIN',
'AGE_15_19_ORIGIN',
'AGE_20_24_ORIGIN',
'AGE_25_34_ORIGIN',
'AGE_35_44_ORIGIN',
'AGE_45_54_ORIGIN',
'AGE_55_59_ORIGIN',
'AGE_60_64_ORIGIN',
'AGE_65_74_ORIGIN',
'AGE_75_84_ORIGIN',
'AGE_85P_ORIGIN',
'MEDIAN_AGE_ORIGIN',
'TOTAL_HHLDS_ORIGIN',
'AVG_HHLD_SIZE_ORIGIN',
'AVG_FAMILY_SIZE_ORIGIN',
'TOTAL_POP_18_24_ORIGIN',
'TOTAL_POP_18_24_NO_HIGH_SCHOOL_ORIGIN',
'TOTAL_POP_18_24_HIGH_SCHOOL_ORIGIN',
'TOTAL_POP_18_24_SOME_COLLEGE_ORIGIN',
'TOTAL_POP_18_24_SOME_BACHELORS_HIGHER_ORIGIN',
'TOTAL_POP_25P_ORIGIN',
'TOTAL_POP_25P_LESS_THAN_9TH_ORIGIN',
'TOTAL_POP_25P_9TH_TO_12TH_ORIGIN',
'TOTAL_POP_25P_HIGH_SCHOOL_ORIGIN',
'TOTAL_POP_25P_SOME_COLLEGE_ORIGIN',
'TOTAL_POP_25P_ASSOCIATES_ORIGIN',
'TOTAL_POP_25P_BACHELORS_ORIGIN',
'TOTAL_POP_25P_GRADUATE_ORIGIN',
'TOTAL_POP_25P_HIGH_SCHOOL_HIGHER_ORIGIN',
'TOTAL_POP_25P_BACHELORS_HIGHER_ORIGIN',
'TOTAL_HHLDS_LESS_10K_ORIGIN',
'TOTAL_HHLDS_10K_15K_ORIGIN',
'TOTAL_HHLDS_15K_25K_ORIGIN',
'TOTAL_HHLDS_25K_35K_ORIGIN',
'TOTAL_HHLDS_35K_50K_ORIGIN',
'TOTAL_HHLDS_50K_75K_ORIGIN',
'TOTAL_HHLDS_75K_100K_ORIGIN',
'TOTAL_HHLDS_100K_150K_ORIGIN',
'TOTAL_HHLDS_150K_200K_ORIGIN',
'TOTAL_HHLDS_200KP_ORIGIN',
'HHLDS_MEDIAN_INCOME_ORIGIN',
'HHLDS_MEAN_INCOME_ORIGIN',
'TOTAL_POP_DESTINATION',
'TOTAL_MALE_DESTINATION',
'TOTAL_FEMALE_DESTINATION',
'AGE_5U_DESTINATION',
'AGE_5_9_DESTINATION',
'AGE_10_14_DESTINATION',
'AGE_15_19_DESTINATION',
'AGE_20_24_DESTINATION',
'AGE_25_34_DESTINATION',
'AGE_35_44_DESTINATION',
'AGE_45_54_DESTINATION',
'AGE_55_59_DESTINATION',
'AGE_60_64_DESTINATION',
'AGE_65_74_DESTINATION',
'AGE_75_84_DESTINATION',
'AGE_85P_DESTINATION',
'MEDIAN_AGE_DESTINATION',
'CENSUS_TRACT_DESTINATION',
'TOTAL_HHLDS_DESTINATION',
'AVG_HHLD_SIZE_DESTINATION',
'AVG_FAMILY_SIZE_DESTINATION',
'TOTAL_POP_18_24_DESTINATION',
'TOTAL_POP_18_24_NO_HIGH_SCHOOL_DESTINATION',
'TOTAL_POP_18_24_HIGH_SCHOOL_DESTINATION',
'TOTAL_POP_18_24_SOME_COLLEGE_DESTINATION',
'TOTAL_POP_18_24_SOME_BACHELORS_HIGHER_DESTINATION',
'TOTAL_POP_25P_DESTINATION',
'TOTAL_POP_25P_LESS_THAN_9TH_DESTINATION',
'TOTAL_POP_25P_9TH_TO_12TH_DESTINATION',
'TOTAL_POP_25P_HIGH_SCHOOL_DESTINATION',
'TOTAL_POP_25P_SOME_COLLEGE_DESTINATION',
'TOTAL_POP_25P_ASSOCIATES_DESTINATION',
'TOTAL_POP_25P_BACHELORS_DESTINATION',
'TOTAL_POP_25P_GRADUATE_DESTINATION',
'TOTAL_POP_25P_HIGH_SCHOOL_HIGHER_DESTINATION',
'TOTAL_POP_25P_BACHELORS_HIGHER_DESTINATION',
'TOTAL_HHLDS_LESS_10K_DESTINATION',
'TOTAL_HHLDS_10K_15K_DESTINATION',
'TOTAL_HHLDS_15K_25K_DESTINATION',
'TOTAL_HHLDS_25K_35K_DESTINATION',
'TOTAL_HHLDS_35K_50K_DESTINATION',
'TOTAL_HHLDS_50K_75K_DESTINATION',
'TOTAL_HHLDS_75K_100K_DESTINATION',
'TOTAL_HHLDS_100K_150K_DESTINATION',
'TOTAL_HHLDS_150K_200K_DESTINATION',
'TOTAL_HHLDS_200KP_DESTINATION',
'HHLDS_MEDIAN_INCOME_DESTINATION',
'HHLDS_MEAN_INCOME_DESTINATION',
'ORIGIN_CENSUS_TRACT',
'DESTINATION_CENSUS_TRACT',
'TOTAL_EMP_LEHD_OD',
'29U_EMP_LEHD_OD',
'30_54_EMP_LEHD_OD',
'55P_EMP_LEHD_OD',
'1250U_EMP_LEHD_OD',
'1251_3333_EMP_LEHD_OD',
'3333P_EMP_LEHD_OD',
'TOTAL_EMP_RAC_ORIGIN',
'29U_EMP_RAC_ORIGIN',
'30_54_EMP_RAC_ORIGIN',
'55P_EMP_RAC_ORIGIN',
'1250U_EMP_RAC_ORIGIN',
'1251_3333__RAC_ORIGIN',
'3333P_EMP_RAC_ORIGIN',
'FOOD_SERVICE_EMP_RAC_ORIGIN',
'TOTAL_EMP_RAC_DESTINATION',
'29U_EMP_RAC_DESTINATION',
'30_54_EMP_RAC_DESTINATION',
'55P_EMP_RAC_DESTINATION',
'1250U_EMP_RAC_DESTINATION',
'1251_3333_EMP_RAC_DESTINATION',
'3333P_EMP_RAC_DESTINATION',
'FOOD_SERVICE_EMP_RAC_DESTINATION',
'TOTAL_EMP_WAC_ORIGIN',
'29U_EMP_WAC_ORIGIN',
'30_54_EMP_WAC_ORIGIN',
'55P_EMP_WAC_ORIGIN',
'1250U_EMP_WAC_ORIGIN',
'1251_3333__WAC_ORIGIN',
'333P_EMP_WAC_ORIGIN',
'FOOD_SERVICE_EMP_WAC_ORIGIN',
'TOTAL_EMP_WAC_DESTINATION',
'29U_EMP_WAC_DESTINATION',
'30_54_EMP_WAC_DESTINATION',
'55P_EMP_WAC_DESTINATION',
'1250U_EMP_WAC_DESTINATION',
'1251_3333__WAC_DESTINATION',
'3333P_EMP_WAC_DESTINATION',
'FOOD_SERVICE_EMP_WAC_DESTINATION']

est.columns = rename

est.head()

drop_str = ['ORIGIN',
'DESTINATION',
'MONTH',
'CENSUS_TRACT_DESTINATION','TRACTCE10_PICKUP','TRACTCE10_DEST','TRACTCE10_RAC_ORIGIN','TRACTCE10_RAC_DESTINATION','TRACTCE10_WAC_ORIGIN','TRACTCE10_WAC_DESTINATION']


est_stats = est.drop(drop_str, axis = 1)

def find_mean(column):
    column = pd.to_numeric(column)
    output = column.mean()
    
    return output

def find_min(column):
    column = pd.to_numeric(column)
    output = column.min()
    
    return output

def find_max(column):
    column = pd.to_numeric(column)
    output = column.max()
    
    return output

est_stats = est_stats.replace('-', '0')
est_stats = est_stats.replace('250,000+', '250000')

df = pd.DataFrame()
df['Mean'] = est_stats.apply(lambda column: find_mean(column), axis = 0)
df['Min'] = est_stats.apply(lambda column: find_min(column), axis = 0)
df['Max'] = est_stats.apply(lambda column: find_max(column), axis = 0)

df.head()

df.to_csv('D:/TNC-Demand-Model/Outputs/Descriptive Statistics of 2019 Estimation File.csv')

first_100 = est.head(100)

first_100.to_csv('D:/TNC-Demand-Model/Outputs/Chi_Estimation_File_2019 (First 100 rows).csv')

cols = est.columns

df2 = pd.DataFrame()
df2['VARIABLE'] = cols

df2.to_csv('D:/TNC-Demand-Model/Outputs/Data Dictionary for Chi Estimation File.csv')
